Refactor/android/double click: 다중 클릭 방지 개선 및 UI, Notification 개선

+++ android/app/src/main/java/com/teameetmeet/meetmeet/presentation/notification/EventNotificationViewModel.kt
@@ -82,12 +82,14 @@ class EventNotificationViewModel @Inject constructor(
 
     fun onDeleteAll() {
         viewModelScope.launch {
-            userRepository.deleteUserNotification(
-                _eventNotificationList.value.map { it.inviteId }.joinToString(",")
-            ).catch {
-                emitExceptionEvent(it, R.string.notification_delete_fail)
-            }.collectLatest {
-                fetchEventNotificationList()
+            if (_eventNotificationList.value.isNotEmpty()) {
+                userRepository.deleteUserNotification(
+                    _eventNotificationList.value.map { it.inviteId }.joinToString(",")
+                ).catch {
+                    emitExceptionEvent(it, R.string.notification_delete_fail)
+                }.collectLatest {
+                    fetchEventNotificationList()
+                }
             }
         }
     }
